0.4.0

+++ block-tlb/src/AddrVar.kt
@@ -75,7 +75,7 @@ private object AddrVarTlbConstructor : TlbConstructor<AddrVar>(
     ): AddrVar = cellSlice {
         val anycast = loadTlb(MaybeAnycast)
         val addrLen = loadUInt(9).toInt()
-        val workchainId = loadInt(32).toInt()
+        val workchainId = loadInt(32)
         val address = loadBits(addrLen)
         AddrVar(anycast, addrLen, workchainId, address)
     }
